Pass tmp file to bsub instead of contents of tmp file
This also requires that the generated tmp file is accessible from the bsub job.

Previously, the contents of the tmp file that starts the spark job were dumped to string as input for the bsub command. For tmp files that contain a lot of (application) parameters, this would break the monitoring system.